# lib/api-zod/src/generated/types/filmInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FilmInfo {
  /** Whether a film was identified */
  found: boolean;
  /**
   * Film title
   * @nullable
   */
  title?: string | null;
  /**
   * Year of release
   * @nullable
   */
  year?: number | null;
  /**
   * Director name(s)
   * @nullable
   */
  director?: string | null;
  /**
   * Film genre
   * @nullable
   */
  genre?: string | null;
  /**
   * Brief description of the film
   * @nullable
   */
  description?: string | null;
  /**
   * Confidence level of the identification
   * @nullable
   */
  confidence?: string | null;
  /**
   * Description of the identified scene
   * @nullable
   */
  sceneDescription?: string | null;
}