# lib/api-zod/src/generated/types/identifySceneRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IdentifySceneRequest {
  /** Base64-encoded image data (without data URI prefix) */
  imageBase64: string;
  /** MIME type of the image (e.g. image/jpeg, image/png) */
  mimeType: string;
}